requirements.txt
django_location_field==2.1.0
Django==3.0.6
Pillow==7.1.2
